lib.sh: post_login: don't test -z on $AUTH anymore
We have 2 stages of error:
1. Returned by post_login: "login request failed" means that HTTP POST query has not been sucessful (no cookie or no result)
Example1: rapidshare upload collectorszone unknown login.
Example2: megaupload premium download
2. Detected by module: "login process failed" means that login/password is wrong
Example: rapidshare upload collectorszone good login but bad password

Be careful again (post_login) most website are returning a cookie even if login/password is wrong;
module must check post_login (stdout) result.

megaupload:
- BASEURL definition was missing in megaupload_upload()

letitbit:
- drop unused post_login code, we do support $AUTH yet

x7_to:
- post_login call update

rapidshare:
- post_login call update
- fix rapidshare_upload_freezone() when password is wrong

git-svn-id: https://plowshare.googlecode.com/svn/trunk@992 51fab9ec-ea3e-11dd-80ed-bda8ba2650cd

+++ src/modules/megaupload.sh
@@ -48,17 +48,17 @@ megaupload_download() {
     COOKIES=$(create_tempfile)
     ERRORURL="http://www.megaupload.com/?c=msg"
     URL=$(echo "$1" | replace 'rotic\.com/' 'porn\.com/')
-    BASEURL=$(basename_url "$URL")
 
     # Try to login (if $AUTH not null)
-    LOGIN_DATA='login=1&redir=1&username=$USER&password=$PASSWORD'
-    test "$AUTH" && { 
+    if [ -n "$AUTH" ]; then
+        LOGIN_DATA='login=1&redir=1&username=$USER&password=$PASSWORD'
+        BASEURL=$(basename_url "$URL")
         post_login "$AUTH" "$COOKIES" "$LOGIN_DATA" "$BASEURL/?c=login" >/dev/null || {
-            log_error "login process failed"
             rm -f $COOKIES
             return 1
         }
-    }
+    fi
+
     echo $URL | grep -q "\.com/?d=" ||
         URL=$(curl -I "$URL" | grep_http_header_location)
 
@@ -176,14 +176,15 @@ megaupload_upload() {
     local FILE=$1
     local DESTFILE=${2:-$FILE}
     local LOGINURL="http://www.megaupload.com/?c=login"
+    local BASEURL=$(basename_url "$LOGINURL")
 
     COOKIES=$(create_tempfile)
-    LOGIN_DATA='login=1&redir=1&username=$USER&password=$PASSWORD'
-    post_login "$AUTH" "$COOKIES" "$LOGIN_DATA" "$LOGINURL" >/dev/null
-    if [ "$?" -eq 1 ]; then
-        log_error "login process failed"
-        rm -f $COOKIES
-        return 1
+    if [ -n "$AUTH" ]; then
+        LOGIN_DATA='login=1&redir=1&username=$USER&password=$PASSWORD'
+        post_login "$AUTH" "$COOKIES" "$LOGIN_DATA" "$LOGINURL" >/dev/null || {
+            rm -f $COOKIES
+            return 1
+        }
     fi
 
     if [ "$MULTIFETCH" ]; then
@@ -191,8 +192,8 @@ megaupload_upload() {
       STATUSURL="http://www.megaupload.com/?c=multifetch&s=transferstatus"
       STATUSLOOPTIME=5
 
-      # FIXME: this is not good, cookie will contain sessionid
-      [ -z "$COOKIES" ] &&
+      # Cookie file must contain sessionid
+      [ -s "$COOKIES" ] ||
           { log_error "Premium account required to use multifetch"; return 2; }
 
       log_debug "spawn URL fetch process: $FILE"
@@ -261,12 +262,10 @@ megaupload_delete() {
 
     COOKIES=$(create_tempfile)
     LOGIN_DATA='login=1&redir=1&username=$USER&password=$PASSWORD'
-    post_login "$AUTH" "$COOKIES" "$LOGIN_DATA" "$LOGINURL" >/dev/null
-    if [ "$?" -eq 1 ]; then
-        log_error "login process failed"
+    post_login "$AUTH" "$COOKIES" "$LOGIN_DATA" "$LOGINURL" >/dev/null || {
         rm -f $COOKIES
         return 1
-    fi
+    }
 
     FILEID=$(echo "$URL" | parse "." "d=\(.*\)")
     DATA="action=deleteItems&items_list[]=file_$FILEID&mode=modeAll&parent_id=0"
@@ -284,13 +283,14 @@ megaupload_list() {
     URL=$1
     XMLURL="http://www.megaupload.com/xml/folderfiles.php"
     FOLDERID=$(echo "$URL" | grep -o "f=[^=]\+$" | cut -d"=" -f2-) ||
-      { log_error "cannot parse url: $URL"; return 1; }
+        { log_error "cannot parse url: $URL"; return 1; }
     XML=$(curl "$XMLURL/?folderid=$FOLDERID")
     match "<FILES></FILES>" "$XML" &&
-      { log_notice "Folder link not found"; return 254; }
+        { log_notice "Folder link not found"; return 254; }
     echo "$XML" | parse_all_attr "<ROW" "url"
 }
 
+# FIXME: get rid of this
 get_location() {
     grep "^[Ll]ocation:" | head -n1 | cut -d":" -f2- | xargs
 }
